--- src/Lingya.Pagination/SearchMethod.cs
@@ -1,18 +1,0 @@
-namespace Lingya.Pagination
-{
-    public enum SearchMethod
-    {
-        /// <summary>
-        /// 包含
-        /// </summary>
-        Contains,
-        /// <summary>
-        /// 开始字符相同
-        /// </summary>
-        StartsWith,
-        /// <summary>
-        /// 结束字符相同
-        /// </summary>
-        EndsWith
-    }
-}
